修复：精度丢失

+++ src/Application.Core/net/server/channel/handlers/SpecialMoveHandler.cs
@@ -120,7 +120,7 @@ public class SpecialMoveHandler : AbstractPacketHandler
             StatEffect ef = s.getEffect(chr.getSkillLevel(s));
 
             int lose = chr.safeAddHP(-1 * (chr.getCurrentMaxHp() / ef.getX()));
-            int gain = -lose * (ef.getY() / 100);
+            int gain = (int)(-lose * (ef.getY() / 100.0));
             chr.addMP(gain);
         }
         else if (skillid == SuperGM.HEAL_PLUS_DISPEL)
